+++ src/ProgramBuilder.ts
@@ -82,7 +82,7 @@ export class ProgramBuilder {
     public getDiagnostics() {
         return [
             ...this.staticDiagnostics,
-            ...this.program ? this.program.getDiagnostics() : []
+            ...(this.program?.getDiagnostics() ?? [])
         ];
     }
 
@@ -115,7 +115,6 @@ export class ProgramBuilder {
         this.program = this.createProgram();
 
         //parse every file in the entire project
-        this.logger.log('Parsing files');
         await this.loadAllFilesAST();
 
         if (this.options.watch) {
@@ -249,11 +248,13 @@ export class ProgramBuilder {
         return runPromise;
     }
 
-    private printDiagnostics() {
+    private printDiagnostics(diagnostics?: BsDiagnostic[]) {
         if (this.options.showDiagnosticsInConsole === false) {
             return;
         }
-        let diagnostics = this.getDiagnostics();
+        if (!diagnostics) {
+            diagnostics = this.getDiagnostics();
+        }
 
         //group the diagnostics by file
         let diagnosticsByFile = {} as Record<string, BsDiagnostic[]>;
@@ -309,7 +310,6 @@ export class ProgramBuilder {
             if (cancellationToken.isCanceled === true) {
                 return -1;
             }
-            this.logger.log('Validating project');
             //validate program
             this.validateProject();
 
@@ -318,9 +318,10 @@ export class ProgramBuilder {
                 return -1;
             }
 
-            this.printDiagnostics();
+            const diagnostics = this.getDiagnostics();
+            this.printDiagnostics(diagnostics);
             wereDiagnosticsPrinted = true;
-            let errorCount = this.getDiagnostics().filter(x => x.severity === DiagnosticSeverity.Error).length;
+            let errorCount = diagnostics.filter(x => x.severity === DiagnosticSeverity.Error).length;
 
             if (errorCount > 0) {
                 this.logger.log(`Found ${errorCount} ${errorCount === 1 ? 'error' : 'errors'}`);
@@ -427,59 +428,61 @@ export class ProgramBuilder {
      * Parse and load the AST for every file in the project
      */
     private async loadAllFilesAST() {
-        let errorCount = 0;
-        let files = await this.logger.time(LogLevel.debug, ['ProgramBuilder.loadAllFilesAST()'], async () => {
-            return util.getFilePaths(this.options);
-        });
-        this.logger.debug('ProgramBuilder.loadAllFilesAST() files:', files);
+        await this.logger.time(LogLevel.log, ['Parsing files'], async () => {
+            let errorCount = 0;
+            let files = await this.logger.time(LogLevel.debug, ['getFlePaths'], async () => {
+                return util.getFilePaths(this.options);
+            });
+            this.logger.trace('ProgramBuilder.loadAllFilesAST() files:', files);
 
-        const typedefFiles = [] as FileObj[];
-        const nonTypedefFiles = [] as FileObj[];
-        for (const file of files) {
-            const srcLower = file.src.toLowerCase();
-            if (srcLower.endsWith('.d.bs')) {
-                typedefFiles.push(file);
-            } else {
-                nonTypedefFiles.push(file);
-            }
-        }
-
-        //preload every type definition file first, which eliminates duplicate file loading
-        await Promise.all(
-            typedefFiles.map(async (fileObj) => {
-                try {
-                    this.program.addOrReplaceFile(
-                        fileObj,
-                        await this.getFileContents(fileObj.src)
-                    );
-                } catch (e) {
-                    //log the error, but don't fail this process because the file might be fixable later
-                    this.logger.log(e);
+            const typedefFiles = [] as FileObj[];
+            const nonTypedefFiles = [] as FileObj[];
+            for (const file of files) {
+                const srcLower = file.src.toLowerCase();
+                if (srcLower.endsWith('.d.bs')) {
+                    typedefFiles.push(file);
+                } else {
+                    nonTypedefFiles.push(file);
                 }
-            })
-        );
+            }
 
-        const acceptableExtensions = ['.bs', '.brs', '.xml'];
-        //parse every file other than the type definitions
-        await Promise.all(
-            nonTypedefFiles.map(async (fileObj) => {
-                try {
-                    let fileExtension = path.extname(fileObj.src).toLowerCase();
-
-                    //only process certain file types
-                    if (acceptableExtensions.includes(fileExtension)) {
+            //preload every type definition file first, which eliminates duplicate file loading
+            await Promise.all(
+                typedefFiles.map(async (fileObj) => {
+                    try {
                         this.program.addOrReplaceFile(
                             fileObj,
                             await this.getFileContents(fileObj.src)
                         );
+                    } catch (e) {
+                        //log the error, but don't fail this process because the file might be fixable later
+                        this.logger.log(e);
                     }
-                } catch (e) {
-                    //log the error, but don't fail this process because the file might be fixable later
-                    this.logger.log(e);
-                }
-            })
-        );
-        return errorCount;
+                })
+            );
+
+            const acceptableExtensions = ['.bs', '.brs', '.xml'];
+            //parse every file other than the type definitions
+            await Promise.all(
+                nonTypedefFiles.map(async (fileObj) => {
+                    try {
+                        let fileExtension = path.extname(fileObj.src).toLowerCase();
+
+                        //only process certain file types
+                        if (acceptableExtensions.includes(fileExtension)) {
+                            this.program.addOrReplaceFile(
+                                fileObj,
+                                await this.getFileContents(fileObj.src)
+                            );
+                        }
+                    } catch (e) {
+                        //log the error, but don't fail this process because the file might be fixable later
+                        this.logger.log(e);
+                    }
+                })
+            );
+            return errorCount;
+        });
     }
 
     /**