+++ src/parser/tests/expression/TernaryExpression.spec.ts
@@ -266,6 +266,14 @@ describe('ternary expressions', () => {
             program.dispose();
         });
 
+        it('uses the proper prefix when aliased package is installed', () => {
+            program.addOrReplaceFile('source/roku_modules/rokucommunity_bslib/bslib.brs', '');
+            testTranspile(
+                `a = user = invalid ? "no user" : "logged in"`,
+                `a = rokucommunity_bslib_ternary(user = invalid, "no user", "logged in")`
+            );
+        });
+
         it('simple consequents', () => {
             testTranspile(
                 `a = user = invalid ? "no user" : "logged in"`,