+++ src/util.spec.ts
@@ -630,4 +630,16 @@ describe('util', () => {
             expect(stub.callCount).to.equal(0);
         });
     });
+
+    describe('copyBslibToStaging', () => {
+        it('copies from local bslib dependency', async () => {
+            await util.copyBslibToStaging(tempDir);
+            expect(fsExtra.pathExistsSync(`${tempDir}/source/bslib.brs`)).to.be.true;
+            expect(
+                /^function bslib_toString\(/mg.exec(
+                    fsExtra.readFileSync(`${tempDir}/source/bslib.brs`).toString()
+                )
+            ).not.to.be.null;
+        });
+    });
 });