+++ src/util.ts
@@ -1175,6 +1175,32 @@ export class Util {
             range: attr.range
         } as SGAttribute;
     }
+
+    /**
+     * Copy the version of bslib from local node_modules to the staging folder
+     */
+    public async copyBslibToStaging(stagingDir: string) {
+        //copy bslib to the output directory
+        await fsExtra.ensureDir(standardizePath(`${stagingDir}/source`));
+        // eslint-disable-next-line
+        const bslib = require('@rokucommunity/bslib');
+        let source = bslib.source as string;
+
+        //apply the `bslib_` prefix to the functions
+        let match: RegExpExecArray;
+        const positions = [] as number[];
+        const regexp = /^(\s*(?:function|sub)\s+)([a-z0-9_]+)/mg;
+        // eslint-disable-next-line no-cond-assign
+        while (match = regexp.exec(source)) {
+            positions.push(match.index + match[1].length);
+        }
+
+        for (let i = positions.length - 1; i >= 0; i--) {
+            const position = positions[i];
+            source = source.slice(0, position) + 'bslib_' + source.slice(position);
+        }
+        await fsExtra.writeFile(`${stagingDir}/source/bslib.brs`, source);
+    }
 }
 
 /**